+++ psiflow/utils.py
@@ -17,6 +17,17 @@ from parsl.config import Config
 
 
 @typeguard.typechecked
+def _create_if_empty(outputs: List[File] = []) -> None:
+    try:
+        with open(inputs[1], 'r') as f:
+            f.read()
+    except FileNotFoundError: # create it if it doesn't exist
+        with open(inputs[1], 'w+') as f:
+            f.write('')
+create_if_empty = python_app(_create_if_empty, executors=['default'])
+
+
+@typeguard.typechecked
 def get_parsl_config_from_file(
         path_config: Union[Path, str],
         path_internal: Union[Path, str],
@@ -54,21 +65,15 @@ def get_index_element_mask(
 
 
 @typeguard.typechecked
-def _new_file(path: Union[Path, str], prefix: str, suffix: str) -> str:
-    _, name = tempfile.mkstemp(
-            suffix=suffix,
-            prefix=prefix,
-            dir=path,
-            )
-    return name
-
-
-@typeguard.typechecked
 def _copy_data_future(inputs: List[File] = [], outputs: List[File] = []) -> None:
     import shutil
+    from pathlib import Path
     assert len(inputs)  == 1
     assert len(outputs) == 1
-    shutil.copyfile(inputs[0], outputs[0])
+    if Path(inputs[0]).is_file():
+        shutil.copyfile(inputs[0], outputs[0])
+    else: # no need to copy empty file
+        pass
 copy_data_future = python_app(_copy_data_future, executors=['default'])
